Add dashboard context
