+++ frontend/src/metabase/dashboard/containers/DashboardApp/use-slow-card-notification.ts
@@ -1,0 +1,69 @@
+import { useCallback, useEffect } from "react";
+import { t } from "ttag";
+
+import { DASHBOARD_SLOW_TIMEOUT } from "metabase/dashboard/constants";
+import { useDashboardContext } from "metabase/dashboard/context";
+import { useLoadingTimer } from "metabase/hooks/use-loading-timer";
+import { useUniqueId } from "metabase/hooks/use-unique-id";
+import { useWebNotification } from "metabase/hooks/use-web-notification";
+import { useDispatch } from "metabase/lib/redux";
+import { addUndo, dismissUndo } from "metabase/redux/undo";
+
+export const useSlowCardNotification = () => {
+  const { dashboard, isRunning, isLoadingComplete } = useDashboardContext();
+
+  const dispatch = useDispatch();
+
+  const { requestPermission, showNotification } = useWebNotification();
+
+  const slowToastId = useUniqueId();
+
+  useEffect(() => {
+    if (isLoadingComplete) {
+      if (
+        "Notification" in window &&
+        Notification.permission === "granted" &&
+        document.hidden
+      ) {
+        showNotification(
+          t`All Set! ${dashboard?.name} is ready.`,
+          t`All questions loaded`,
+        );
+      }
+    }
+
+    return () => {
+      dispatch(dismissUndo({ undoId: slowToastId }));
+    };
+  }, [
+    dashboard?.name,
+    dispatch,
+    isLoadingComplete,
+    showNotification,
+    slowToastId,
+  ]);
+
+  const onConfirmToast = useCallback(async () => {
+    await requestPermission();
+    dispatch(dismissUndo({ undoId: slowToastId }));
+  }, [dispatch, requestPermission, slowToastId]);
+
+  const onTimeout = useCallback(() => {
+    if ("Notification" in window && Notification.permission === "default") {
+      dispatch(
+        addUndo({
+          id: slowToastId,
+          timeout: false,
+          message: t`Would you like to be notified when this dashboard is done loading?`,
+          action: onConfirmToast,
+          actionLabel: t`Turn on`,
+        }),
+      );
+    }
+  }, [dispatch, onConfirmToast, slowToastId]);
+
+  useLoadingTimer(isRunning, {
+    timer: DASHBOARD_SLOW_TIMEOUT,
+    onTimeout,
+  });
+};
