+++ frontend/src/metabase/dashboard/types/embed-display-options.ts
@@ -5,9 +5,9 @@ import type {
 } from "metabase/public/lib/types";
 import type { ClickActionModeGetter } from "metabase/visualizations/types";
 
-type EmbedBackground = boolean;
+export type EmbedBackground = boolean;
 
-type EmbedTitle = boolean;
+export type EmbedTitle = boolean;
 
 export type EmbedHideParameters = string | null;
 export type EmbedHideParametersControls = {
